Add support for retrieval of a jobid for a given test
For some tests, we want to retrieve a specific job ID for a test.
For example, if we want to retrieve a prepared qcow2 image from openQA
to use with libvirt on a local machine.

To support this, add a new API endpoint:
/v1/latest_job/<arch>/<distri>/<flavor>/<version>/<test>

for example:
/v1/latest_job/x86_64/sle/Online/15-SP2/create_hdd_minimal_base+sdk

If successful, the openQA job ID is returned. If no job can be found, an
error 404 is returned.

Signed-off-by: Michael Moese <frankenmichl@posteo.de>